Rename config.rs to config_manager.rs
