+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod config;
+pub mod config_manager;
 pub mod error;
 pub mod network_client;
 pub mod oled_client;
